extract common funtions
