+++ lib/scrapers/discount.js
@@ -1,75 +1,7 @@
 import phantom from 'phantom';
 
-function timeoutPromise(ms, promise) {
-  const timeout = new Promise((resolve, reject) => {
-    const id = setTimeout(() => {
-      clearTimeout(id);
-      reject({ timeout: true });
-    }, ms);
-  });
-
-  return Promise.race([
-    promise,
-    timeout,
-  ]);
-}
-
-function waitUntil(asyncTest, timeout = 10000, interval = 100) {
-  const promise = new Promise((resolve, reject) => {
-    function wait() {
-      asyncTest().then((value) => {
-        if (value === true) {
-          resolve();
-        } else {
-          setTimeout(wait, interval);
-        }
-      }).catch((e) => {
-        console.log('Error found. Rejecting.', e);
-        reject();
-      });
-    }
-    wait();
-  });
-  return timeoutPromise(timeout, promise);
-}
-
-async function waitUntilElementFound(page, elementName) {
-  await waitUntil(() => {
-    return page.evaluate((elementName) => {
-      return document.getElementById(elementName) !== null;
-    }, elementName);
-  });
-}
-
-async function waitForUrl(page, url) {
-  try {
-    await waitUntil(async () => {
-      const current = await page.property('url');
-      return current === url;
-    }, 20000, 1000);
-  } catch (e) {
-    if (e && e.timeout) {
-      const current = await page.property('url');
-      console.log(`timeout reached this page: ${current}`);
-      e.lastUrl = current;
-    }
-    throw e;
-  }
-}
-
-async function fillInput(page, inputName, inputValue) {
-  await page.evaluate((inputName, inputValue) => {
-    const input = document.getElementById(inputName);
-    input.value = inputValue;
-  }, inputName, inputValue);
-}
-
-async function clickButton(page, buttonName) {
-  await page.evaluate((buttonName) => {
-    const button = document.getElementById(buttonName);
-    button.click();
-  }, buttonName);
-}
+import waitForUrl from '../helpers/navigation';
+import { waitUntilElementFound, fillInput, clickButton } from '../helpers/elements-interactions';
 
 async function scrape(credentials) {
   console.log('discount: start scraping');
